requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-mysql==2.4.1
djangorestframework==3.9.1
drf-nested-routers==0.91
idna==2.8
mysqlclient==1.4.1
pytz==2018.9
requests==2.21.0

celery[redis]==4.2.1

urllib3==1.24.1
boto3==1.9.98
zxing==0.9.3
django-filter==2.1.0
django-sns-sqs-services==0.0.5
sentry-sdk==0.7.14
django-service-objects==0.5.0